requirements.txt
django>=5.0
djangorestframework
pandas
numpy
requests
python-dotenv
hypothesis
polyline
